Add CSS Collapsible demo (#518)
