+++ components/demos/Collapsible/css/styles.css
@@ -1,0 +1,49 @@
+@import '@radix-ui/colors/blackA.css';
+@import '@radix-ui/colors/violet.css';
+
+/* reset */
+button {
+  all: unset;
+}
+
+.CollapsibleRoot {
+  width: 300px;
+}
+
+.IconButton {
+  font-family: inherit;
+  border-radius: 100%;
+  height: 25px;
+  width: 25px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  color: var(--violet11);
+  box-shadow: 0 2px 10px var(--blackA7);
+}
+.IconButton[data-state='closed'] {
+  background-color: white;
+}
+.IconButton[data-state='open'] {
+  background-color: var(--violet3);
+}
+.IconButton:hover {
+  background-color: var(--violet3);
+}
+.IconButton:focus {
+  box-shadow: 0 0 0 2px black;
+}
+
+.Text {
+  color: var(--violet11);
+  font-size: 15px;
+  line-height: 25px;
+}
+
+.Repository {
+  background-color: white;
+  border-radius: 4px;
+  margin: 10px 0;
+  padding: 10px;
+  box-shadow: 0 2px 10px var(--blackA7);
+}
