+++ components/demos/index.jsx
@@ -4,7 +4,7 @@ export { default as AlertDialogDemo } from './AlertDialog/stitches';
 export { default as AspectRatioDemo } from './AspectRatio/stitches';
 export { default as AvatarDemo } from './Avatar/stitches';
 export { default as CheckboxDemo } from './Checkbox/stitches';
-export { default as CollapsibleDemo } from './Collapsible';
+export { default as CollapsibleDemo } from './Collapsible/stitches';
 export { default as ContextMenuDemo } from './ContextMenu';
 export { default as DialogDemo } from './Dialog';
 export { default as DropdownMenuDemo } from './DropdownMenu';
